requirements.txt
openai-whisper~=20240930
dotenv~=0.9.9
python-dotenv~=1.1.0
pydub~=0.25.1
python-telegram-bot~=22.1